pyproject.toml

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"


[project]
name = "mlx-video-with-audio"
dynamic = ["version"]
description = "Generate videos with synchronized audio on Apple Silicon using MLX. Supports text-to-video (T2V) and image-to-video (I2V) with LTX-2."
keywords = [
    "mlx",
    "video-generation",
    "audio-generation",
    "apple-silicon",
    "text-to-video",
    "image-to-video",
    "ltx",
    "diffusion",
    "ai",
    "machine-learning",
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mlx>=0.22.0",
    "numpy",
    "safetensors",
    "huggingface_hub",
    "transformers[tokenizers]",
    "tqdm",
    "opencv-python>=4.12.0.88",
    "Pillow>=10.3.0",
    "mlx-vlm",
    "soundfile>=0.12.0",
]
license = {text="MIT"}
authors = [
    {name = "Prince Canuma", email = "prince.gdt@gmail.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

[project.urls]
Homepage = "https://github.com/james-see/mlx-video-with-audio"
Repository = "https://github.com/james-see/mlx-video-with-audio"
Issues = "https://github.com/james-see/mlx-video-with-audio/issues"

[project.scripts]
"mlx_video.generate" = "mlx_video.generate:main"
"mlx_video.generate_av" = "mlx_video.generate_av:main"
"mlx_video.convert" = "mlx_video.convert:main"

[tool.setuptools.packages.find]
include = ["mlx_video*"]

[tool.setuptools.dynamic]
version = {attr = "mlx_video.version.__version__"}

[project.optional-dependencies]
dev = [
    "pytest",
]
